+++ src/qmap/mod.rs
@@ -9,9 +9,11 @@ pub mod parser;
 #[cfg(feature = "std")]
 pub mod parse_result;
 
+pub mod repr_write;
+
 pub use repr::{
-    Alignment, BaseAlignment, Brush, CheckWritable, Edict, Entity, HalfSpace,
-    Point, QuakeMap, Surface, Vec2, Vec3,
+    Alignment, BaseAlignment, Brush, Edict, Entity, HalfSpace, Point, QuakeMap,
+    Surface, Vec2, Vec3,
 };
 
 #[cfg(feature = "std")]
@@ -19,6 +21,11 @@ pub use parser::parse;
 
 #[cfg(feature = "std")]
 pub use parse_result::{LineError, ParseError, ParseResult};
+
+pub use repr_write::{CheckWritable, ValidationResult};
+
+#[cfg(feature = "std")]
+pub use repr_write::{WriteAttempt, WriteError};
 
 // test suites
